+++ personalInfo/src/main/java/web/mediscreen/personalinfo/PersonalInfoApplication.java
@@ -1,4 +1,4 @@
-package web.mediscreen.personalInfo;
+package web.mediscreen.personalinfo;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
