+++ personalInfo/src/main/java/web/mediscreen/personalinfo/exception/ApiExceptionHandler.java
@@ -1,4 +1,4 @@
-package web.mediscreen.personalInfo.exception;
+package web.mediscreen.personalinfo.exception;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
@@ -13,17 +13,6 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
  */
 @RestControllerAdvice
 public class ApiExceptionHandler {
-
-    /**
-     * @param ex Error raised
-     * @return the JSON answer
-     */
-    @ExceptionHandler(PatientException.class)
-    public ResponseEntity<Object> handleApiException(
-    		PatientException ex) {
-    	final ApiError apiError = new ApiError(HttpStatus.NOT_FOUND, ex.getLocalizedMessage(), "Patient not Found");
-    	return new ResponseEntity<>(apiError, new HttpHeaders(), apiError.getStatus());
-    }  
     
     @ExceptionHandler(PatientExistException.class)
     public ResponseEntity<Object> handleApiException(
