+++ personalInfo/src/main/java/web/mediscreen/personalinfo/exception/DbSaveException.java
@@ -1,11 +1,10 @@
-package web.mediscreen.personalInfo.exception;
+package web.mediscreen.personalinfo.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class PatientExistException extends Exception {
-
+public class DbSaveException extends Exception {
     /**
      * 
      */
@@ -16,7 +15,7 @@ public class PatientExistException extends Exception {
 	/**
 	 * @param message to return
 	 */
-	public PatientExistException(String message) {
+	public DbSaveException(String message) {
 		this.message = message;
 	}
