+++ personalInfo/src/main/java/web/mediscreen/personalinfo/exception/PatientExistException.java
@@ -1,22 +1,22 @@
-package web.mediscreen.personalInfo.exception;
+package web.mediscreen.personalinfo.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(HttpStatus.BAD_REQUEST)
-public class PatientException extends Exception {
+public class PatientExistException extends Exception {
 
-	/**
-	 * 
-	 */
-	private static final long serialVersionUID = 1L;
+    /**
+     * 
+     */
+    private static final long serialVersionUID = 1L;
 
 	private final String message;
 	
 	/**
 	 * @param message to return
 	 */
-	public PatientException(String message) {
+	public PatientExistException(String message) {
 		this.message = message;
 	}
