+++ personalInfo/src/main/java/web/mediscreen/personalinfo/validators/FieldsValueMatchValidator.java
@@ -1,4 +1,4 @@
-package web.mediscreen.personalInfo.validators;
+package web.mediscreen.personalinfo.validators;
 
 import javax.validation.ConstraintValidator;
 import javax.validation.ConstraintValidatorContext;
@@ -6,7 +6,7 @@ import javax.validation.ConstraintValidatorContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import web.mediscreen.personalInfo.model.Patient;
+import web.mediscreen.personalinfo.model.Patient;
 
 /**
  * @author Nico
